+++ packages/block-editor/src/components/block-settings-menu-controls/README.md
@@ -8,7 +8,7 @@ Block Settings Menu Controls appear in the block settings dropdown menu when the
 import { BlockSettingsMenuControls } from '@wordress/block-editor';
 import MyButton from './my-toggle-button';
 
-function ReusableBlocksButtons() {
+function ReusableBlocksMenuItems() {
 	return (
 		<BlockSettingsMenuControls>
 			{ ( { onClose } ) => (
