+++ packages/editor/src/components/provider/index.js
@@ -20,7 +20,7 @@ import {
 import apiFetch from '@wordpress/api-fetch';
 import { addQueryArgs } from '@wordpress/url';
 import { decodeEntities } from '@wordpress/html-entities';
-import { ReusableBlocksButtons } from '@wordpress/reusable-blocks';
+import { ReusableBlocksMenuItems } from '@wordpress/reusable-blocks';
 
 /**
  * Internal dependencies
@@ -292,7 +292,7 @@ class EditorProvider extends Component {
 								useSubRegistry={ false }
 							>
 								{ children }
-								<ReusableBlocksButtons />
+								<ReusableBlocksMenuItems />
 								<ConvertToGroupButtons />
 							</BlockEditorProvider>
 						</BlockContextProvider>
